usb: store controller flags in one byte

+++ ARM/Nordic/include/usb_ctrlr.h
@@ -306,7 +306,7 @@ typedef struct __nRF_Usbd_State
 	// Queued ISO buffer lengths; meaningful while the direction is BUSY.
 	uint16_t IsoDmaLen[2];
 	uint8_t IsoBufState;          //!< READY/BUSY state of the ISO DMA buffers.
-	volatile uint32_t Flags;      //!< Controller power/wake state only.
+	volatile uint8_t Flags;       //!< Controller power/wake state only.
 	hCFifo_t hQue;
 	hCFifo_t hEp0Que;
 	// Non-control endpoints 1-8.
